requirements.txt
# VGP Planemo Scripts - Python Dependencies

# Core dependencies
bioblend>=1.0.0
pandas>=1.0.0
pyyaml>=5.1
requests>=2.20.0

# Planemo for workflow execution
planemo>=0.74.0

# Optional: AWS CLI (for GenomeArk data access)
# Install separately: pip install awscli
